# Improve efficiency