--- a/src/gfn/gym/set_addition.py
+++ b/src/gfn/gym/set_addition.py
@@ -28,6 +28,7 @@ class SetAddition(DiscreteEnv):
         max_items: int,
         reward_fn: Callable,
         fixed_length: bool = False,
+        check_action_validity: bool = True,
     ):
         """Initializes the SetAddition environment.
 
@@ -45,7 +46,12 @@ class SetAddition(DiscreteEnv):
         s0 = torch.zeros(n_items)
         state_shape = (n_items,)
 
-        super().__init__(n_actions, s0, state_shape)
+        super().__init__(
+            n_actions,
+            s0,
+            state_shape,
+            check_action_validity=check_action_validity,
+        )
         self.States: type[DiscreteStates] = self.States
 
     def get_states_indices(self, states: DiscreteStates):