--- a/src/gfn/utils/common.py
+++ b/src/gfn/utils/common.py
@@ -184,17 +184,17 @@ def ensure_same_device(device1: torch.device, device2: torch.device) -> None:
     Raises:
         ValueError: If the devices are not the same.
     """
-    if device1 == device2:
+    try:
+        assert device1 == device2
         return
+    except AssertionError:
+        pass
 
     if device1.type != device2.type:
         raise ValueError(f"The devices have different types: {device1}, {device2}")
 
+    # Device indices must differ since the device types are the same.
     index1, index2 = device1.index, device2.index
-
-    # Same type and same index.
-    if index1 == index2:
-        return
 
     # Both have not-None index but they are different.
     if index1 is not None and index2 is not None: