… src/gfn/states.py