- [x] I've read the [.github/CONTRIBUTING.md](.github/CONTRIBUTING.md) file
- [x] My code follows the typing guidelines
- [x] I've added appropriate tests
- [x] I've run pre-commit hooks locally

## Description
Improve code efficiency by:
1) making action validity check optional
2) calculating rewards outside the sampling loop
3) and several minor improvements.

When tested using `python -O tutorials/examples/train_hypergrid_simple.py --no_cuda --ndim 2 --height 8` the runtime reduced **from 20s to 15s**.