@hyeok9855 2025-08-26
@josephdviviano 
When tested using `python tutorials/examples/train_hypergrid_simple.py --ndim 2 --height 32 --n_iterations 2000 --no_cuda`,

- Before this PR: ~46 sec
- After this PR: ~40 sec (~13% faster)

When tested with the additional `-O` (which sets `__debug__==True` and skips all assertions) flag,

- Before this PR: ~41 sec
- After this PR: ~35 sec (~15% faster)

All tested on my MBP (M4PRO).